Move `upsertMessage` and `upsertMessageBulk`
