+++ apps/meteor/app/ui-utils/client/index.ts
@@ -9,7 +9,7 @@ export { messageBox } from './lib/messageBox';
 export { popover } from './lib/popover';
 export { readMessage } from './lib/readMessages';
 export { RoomManager } from './lib/RoomManager';
-export { upsertMessage, RoomHistoryManager } from './lib/RoomHistoryManager';
+export { RoomHistoryManager } from './lib/RoomHistoryManager';
 export { mainReady } from './lib/mainReady';
 export { IframeLogin, iframeLogin } from './lib/IframeLogin';
 export { popout } from './lib/popout';
